+++ src/components/notification/notification.js
@@ -1,5 +1,6 @@
 import { utils } from '../../utils/utils';
 import { Locale } from '../locale/locale';
+import { Environment as env } from '../../utils/environment';
 
 // Settings and Options
 const COMPONENT_NAME = 'notification';
@@ -125,7 +126,9 @@ Notification.prototype = {
     const textEl = $(this.notificationEl).find('.notification-text')[0];
     const tooltipApi = $(this.notificationEl).data('tooltip');
 
-    if (textEl.offsetHeight < textEl.scrollHeight || textEl.offsetWidth < textEl.scrollWidth) {
+    // don't show tooltip if browser is safari, safari pops out a tooltip after 3 seconds for overflowing text
+    if (!env.browser.isSafari() &&
+      (textEl.offsetHeight < textEl.scrollHeight || textEl.offsetWidth < textEl.scrollWidth)) {
       if (!tooltipApi) {
         $(this.notificationEl).tooltip({ content: this.settings.message });
       } else {